+++ packages/aws-sdk-signers/src/aws_sdk_signers/interfaces/http.py
@@ -5,29 +5,18 @@ from __future__ import annotations
 
 from collections import OrderedDict
 from collections.abc import AsyncIterable, Iterable, Iterator
-from enum import Enum
-from typing import Protocol, runtime_checkable
+from typing import Literal, Protocol, runtime_checkable
 
+FieldPosition = Literal["header", "trailer"]
+"""The type of a field.
 
-class FieldPosition(Enum):
-    """The type of a field.
+Defines its placement in a request or response.
 
-    Defines its placement in a request or response.
-    """
+header: Header field. In HTTP this is a header as defined in RFC 9110 Section 6.3.
+trailer: Trailer field. In HTTP this is a trailer as defined in RFC 9110 Section 6.5.
 
-    HEADER = 0
-    """Header field.
-
-    In HTTP this is a header as defined in RFC 9110 Section 6.3. Implementations of
-    other protocols may use this FieldPosition for similar types of metadata.
-    """
-
-    TRAILER = 1
-    """Trailer field.
-
-    In HTTP this is a trailer as defined in RFC 9110 Section 6.5. Implementations of
-    other protocols may use this FieldPosition for similar types of metadata.
-    """
+Implementations of other protocols may use this FieldPosition for similar types of metadata.
+"""
 
 
 class Field(Protocol):
@@ -43,7 +32,7 @@ class Field(Protocol):
 
     name: str
     values: list[str]
-    kind: FieldPosition = FieldPosition.HEADER
+    kind: FieldPosition = "header"
 
     def add(self, value: str) -> None:
         """Append a value to a field."""